Services/requirements.txt
Flask
grpcio
grpcio-tools
pika
kafka-python==1.4.3